requirements-dev.txt
-r requirements.txt
pytest==7.2.2
pytest-timeout==2.2.0
pre-commit==3.0.4
transformers==4.38.0
pandas==2.0.3
Pillow==10.3.0
pydantic==2.8.2
